add profiler utils
